feat: stat 부분 클래스로 분리
